requirements.txt
apache-airflow==2.8.1
mlflow==2.10.2
scikit-learn==1.4.0
joblib==1.3.2
numpy==1.26.4
pandas==2.2.0
gunicorn==21.2.0
waitress==3.0.0